+++ packages/eslint-plugin/src/rules/func-call-spacing.ts
@@ -24,7 +24,7 @@ export default createRule<Options, MessageIds>({
   name: 'func-call-spacing',
   meta: {
     deprecated: true,
-    replacedBy: ['@stylistic/func-call-spacing'],
+    replacedBy: ['@stylistic/ts/func-call-spacing'],
     type: 'layout',
     docs: {
       description:
@@ -82,8 +82,7 @@ export default createRule<Options, MessageIds>({
 
     /**
      * Check if open space is present in a function name
-     * @param {ASTNode} node node to evaluate
-     * @returns {void}
+     * @param node node to evaluate
      * @private
      */
     function checkSpacing(