+++ packages/eslint-plugin/src/rules/padding-line-between-statements.ts
@@ -182,8 +182,8 @@ function isCJSRequire(node: TSESTree.Node): boolean {
 /**
  * Checks whether the given node is a block-like statement.
  * This checks the last token of the node is the closing brace of a block.
- * @param sourceCode The source code to get tokens.
  * @param node The node to check.
+ * @param sourceCode The source code to get tokens.
  * @returns `true` if the node is a block-like statement.
  * @private
  */
@@ -323,8 +323,8 @@ function isExpression(
  *
  *     foo()
  *     ;[1, 2, 3].forEach(bar)
- * @param sourceCode The source code to get tokens.
  * @param node The node to get.
+ * @param sourceCode The source code to get tokens.
  * @returns The actual last token.
  * @private
  */
@@ -476,7 +476,7 @@ function verifyForAlways(
             }
             return true;
           },
-        }) || nextNode;
+        }) ?? nextNode;
       const insertText = isTokenOnSameLine(prevToken, nextToken)
         ? '\n\n'
         : '\n';
@@ -591,7 +591,7 @@ export default createRule<Options, MessageIds>({
   name: 'padding-line-between-statements',
   meta: {
     deprecated: true,
-    replacedBy: ['@stylistic/padding-line-between-statements'],
+    replacedBy: ['@stylistic/ts/padding-line-between-statements'],
     type: 'layout',
     docs: {
       description: 'Require or disallow padding lines between statements',