+++ packages/eslint-plugin/src/rules/space-before-function-paren.ts
@@ -21,7 +21,7 @@ export default createRule<Options, MessageIds>({
   name: 'space-before-function-paren',
   meta: {
     deprecated: true,
-    replacedBy: ['@stylistic/space-before-function-paren'],
+    replacedBy: ['@stylistic/ts/space-before-function-paren'],
     type: 'layout',
     docs: {
       description: 'Enforce consistent spacing before function parenthesis',
@@ -70,8 +70,8 @@ export default createRule<Options, MessageIds>({
 
     /**
      * Determines whether a function has a name.
-     * @param {ASTNode} node The function node.
-     * @returns {boolean} Whether the function has a name.
+     * @param node The function node.
+     * @returns Whether the function has a name.
      */
     function isNamedFunction(
       node:
@@ -97,8 +97,7 @@ export default createRule<Options, MessageIds>({
 
     /**
      * Gets the config for a given function
-     * @param {ASTNode} node The function node
-     * @returns {string} "always", "never", or "ignore"
+     * @param node The function node
      */
     function getConfigForFunction(
       node:
@@ -129,8 +128,7 @@ export default createRule<Options, MessageIds>({
 
     /**
      * Checks the parens of a function node
-     * @param {ASTNode} node A function node
-     * @returns {void}
+     * @param node A function node
      */
     function checkFunction(
       node: