+++ packages/typescript-estree/src/parser.ts
@@ -53,6 +53,7 @@ function getProgramAndAST(
     const fromProjectService = useProgramFromProjectService(
       parseSettings.EXPERIMENTAL_projectService,
       parseSettings,
+      hasFullTypeInformation,
     );
     if (fromProjectService) {
       return fromProjectService;
@@ -243,7 +244,7 @@ function parseAndGenerateServices<T extends TSESTreeOptions = TSESTreeOptions>(
     options.filePath &&
     parseAndGenerateServicesCalls[options.filePath] > 1
       ? createIsolatedProgram(parseSettings)
-      : getProgramAndAST(parseSettings, hasFullTypeInformation)!;
+      : getProgramAndAST(parseSettings, hasFullTypeInformation);
 
   /**
    * Convert the TypeScript AST to an ESTree-compatible one, and optionally preserve